[Add] Mesh creation.

+++ VkEngine/Include/Rendering/Vertex.h
@@ -2,7 +2,7 @@
 
 struct Vertex final
 {
-	typedef uint8_t Index;
+	typedef uint16_t Index;
 
 	glm::vec3 position{0};
 	glm::vec3 normal{0, 0, 1};
